Update golangci-lint version and timeout
The version of `golangci-lint` that we are currently using is three
years old. This patch updates it to the latest available version.

In addition the default timeout is one minute, and that isn't enough
when it runs without a cache (which is always the case in the CI jobs):

```
$ golangci-lint cache clean
$ make lint
...
ERRO Running error: context loading failed: failed to load packages: timed out to load packages: context deadline exceeded
ERRO Timeout exceeded: try increasing it by passing --timeout option
```

Increasing the timeout to five minutes allows the execution to finish,
and it takes approximately two minutes:

```
$ golangci-lint cache clean
$ golangci-lint run -v --timeout=5m
...
INFO Execution took 1m47.572011137s
```

This patch increases the timeout to five minutes.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>