MGMT-15165: fix indication on update-hosts script
The current indication of checking cluster's status
is not good enough, mainly for multi nodes scenarios.
E.g. since a cluster ready for installation doesn't
mean that all workers been discovered.
Hence, changed the logic to rely on the exact number
of required nodes, as specified in the
start-cluster-installation script.