Fix UEFI boot on baremetal machines
Due to an issue in GRUB that has been recently
fixed only in fedora[*], boot fails on kernel panic
in some machines. The exact failure is an 'out of memory'
that is thrown since we load the full rootfs.img to memory
using initrd.

Hence, till the fix is backported to rhcos, we can manually
override the EFI dir during the appliance disk image build.

[*] https://github.com/rhboot/grub2/commit/a048b2280ae6a2cf90fd5d2960823843e17e4d66