MGMT-15359: create a virtual device for coreos-installer
We need a solution for the limitation in coreos-installer
that can't install on a device with mounted partitions.

Currently, we are loading the rootfs image to memory on boot,
thus, no partitions are mounted on installation device.
However, this workaround works only with UEFI when secure boot
an TPM are disabled. Also, it isn't supported for some baremetals
when using an older GRUB version.

As an alternative solution to mitigate those issues, we are now
creating a multipath virtual device instead. As we build the
device without the mounted partitions, it can be passed to
coreos-installer and the installation is successfully completed.

The following changes were made to support this solution:
* Download raw variant of CoreOS base image (instead of qcow2)
  * Required for using go-diskfs to fetch partitions info
* Added create-virtual-device.sh script for generating the
  partition table and build the device.
  * Using partitions start-sector and size fetched with go-diskfs.
* Removed old workaround of overriding the EFI folder.

Note: this change depends on supporting the multipath device
      in assisted-installer-agent.