MGMT-15520: Appliance deployment bootable ISO
Added a command for building a deployment ISO
that handles disk image resize and clone into
a target device.

The ISO can be used as an alternative for manually cloning
the appliance disk image with a 'dd' or 'virt-resize'.
I.e. post boot, the ISO automatically invokes 'virt-resize'
command that clones the included appliance.raw file
into a specified target device.

For more details, see 'Deployment ISO' section in the user-guide.