Calculate root partition size precisely
The root partition must be precisely clacluated now
as we manually create it in guestfish script.
I.e. before supporting a compact image, the root partition
was resized automatically. But now, we are giving exact
start and end sectors. Thus, the partition size can be calculated
according to the disk size and its content.
Without this fix, the root partation size won't expand from 8GiB
when specifying a disk size.