AGENT-1193: Allow access to release info when self-signed cert is used
Adds --insecure=true to "oc adm release" commands when
fetching version and digest.

This allows appliance to work with registries that are using
self-signed certificates.

The alterative would be to pass the self-signed certificates
to the appliance when it is operated within a container and
then use the --certificate-authority flag when executing
"oc adm release".